refactoring code

+++ my_app/reward-acceptance-tests/src/main/java/org/javaguru/rewardapp/employee/EmployeeDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -17,6 +19,6 @@ public class EmployeeDTO {
 
     private String lastName;
 
-    private Double bonusCoefficient;
+    private BigDecimal bonusCoefficient;
 
 }
